LPS-64166 move to correct location
