+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-data-provider-web/src/main/java/com/liferay/dynamic/data/mapping/data/provider/web/exportimport/staged/model/repository/DDMDataProviderInstanceStagedModelRepository.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under
@@ -12,27 +12,24 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.data.provider.rest.internal.lar;
+package com.liferay.dynamic.data.mapping.data.provider.web.exportimport.staged.model.repository;
 
-import com.liferay.dynamic.data.mapping.data.provider.DDMDataProvider;
-import com.liferay.dynamic.data.mapping.data.provider.DDMDataProviderTracker;
 import com.liferay.dynamic.data.mapping.data.provider.rest.DDMRESTDataProviderSettings;
-import com.liferay.dynamic.data.mapping.data.provider.web.constants.DDMDataProviderPortletKeys;
 import com.liferay.dynamic.data.mapping.io.DDMFormValuesJSONDeserializer;
 import com.liferay.dynamic.data.mapping.model.DDMDataProviderInstance;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.service.DDMDataProviderInstanceLocalService;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
-import com.liferay.exportimport.kernel.lar.BaseStagedModelDataHandler;
-import com.liferay.exportimport.kernel.lar.ExportImportPathUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
-import com.liferay.exportimport.kernel.lar.StagedModelDataHandler;
+import com.liferay.exportimport.kernel.lar.PortletDataException;
 import com.liferay.exportimport.kernel.lar.StagedModelModifiedDateComparator;
+import com.liferay.exportimport.staged.model.repository.StagedModelRepository;
+import com.liferay.exportimport.staged.model.repository.base.BaseStagedModelRepository;
+import com.liferay.portal.kernel.dao.orm.ExportActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.xml.Element;
 
 import java.util.List;
 
@@ -44,14 +41,42 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(
 	immediate = true,
-	property = {"javax.portlet.name=" + DDMDataProviderPortletKeys.DYNAMIC_DATA_MAPPING_DATA_PROVIDER},
-	service = StagedModelDataHandler.class
+	property = {"model.class.name=com.liferay.dynamic.data.mapping.model.DDMDataProviderInstance"},
+	service = StagedModelRepository.class
 )
-public class DDMDataProviderInstanceStagedModelDataHandler
-	extends BaseStagedModelDataHandler<DDMDataProviderInstance> {
+public class DDMDataProviderInstanceStagedModelRepository
+	extends BaseStagedModelRepository<DDMDataProviderInstance> {
 
-	public static final String[] CLASS_NAMES =
-		{DDMDataProviderInstance.class.getName()};
+	@Override
+	public DDMDataProviderInstance addStagedModel(
+			PortletDataContext portletDataContext,
+			DDMDataProviderInstance dataProviderInstance)
+		throws PortalException {
+
+		long userId = portletDataContext.getUserId(
+			dataProviderInstance.getUserUuid());
+
+		ServiceContext serviceContext = portletDataContext.createServiceContext(
+			dataProviderInstance);
+
+		if (portletDataContext.isDataStrategyMirror()) {
+			serviceContext.setUuid(dataProviderInstance.getUuid());
+		}
+
+		String definition = dataProviderInstance.getDefinition();
+
+		DDMForm ddmForm = DDMFormFactory.create(
+			DDMRESTDataProviderSettings.class);
+
+		DDMFormValues ddmFormValues =
+			_ddmFormValuesJSONDeserializer.deserialize(ddmForm, definition);
+
+		return _ddmDataProviderInstanceLocalService.addDataProviderInstance(
+			userId, dataProviderInstance.getGroupId(),
+			dataProviderInstance.getNameMap(),
+			dataProviderInstance.getDescriptionMap(), ddmFormValues,
+			dataProviderInstance.getType(), serviceContext);
+	}
 
 	@Override
 	public void deleteStagedModel(DDMDataProviderInstance dataProviderInstance)
@@ -67,11 +92,21 @@ public class DDMDataProviderInstanceStagedModelDataHandler
 		throws PortalException {
 
 		DDMDataProviderInstance dataProviderInstance =
-			fetchStagedModelByUuidAndGroupId(uuid, groupId);
+			_ddmDataProviderInstanceLocalService.
+				fetchDDMDataProviderInstanceByUuidAndGroupId(uuid, groupId);
 
 		if (dataProviderInstance != null) {
 			deleteStagedModel(dataProviderInstance);
 		}
+	}
+
+	@Override
+	public void deleteStagedModels(PortletDataContext portletDataContext)
+		throws PortalException {
+
+		_ddmDataProviderInstanceLocalService.deleteDataProviderInstances(
+			portletDataContext.getCompanyId(),
+			portletDataContext.getScopeGroupId());
 	}
 
 	@Override
@@ -94,59 +129,37 @@ public class DDMDataProviderInstanceStagedModelDataHandler
 	}
 
 	@Override
-	public String[] getClassNames() {
-		return CLASS_NAMES;
+	public ExportActionableDynamicQuery getExportActionableDynamicQuery(
+		PortletDataContext portletDataContext) {
+
+		return _ddmDataProviderInstanceLocalService.
+			getExportActionableDynamicQuery(portletDataContext);
 	}
 
 	@Override
-	public String getDisplayName(DDMDataProviderInstance dataProviderInstance) {
-		return dataProviderInstance.getNameCurrentValue();
-	}
-
-	@Override
-	protected void doExportStagedModel(
+	public void restoreStagedModel(
 			PortletDataContext portletDataContext,
 			DDMDataProviderInstance dataProviderInstance)
-		throws Exception {
-
-		Element dataProviderInstanceElement =
-			portletDataContext.getExportDataElement(dataProviderInstance);
-
-		portletDataContext.addClassedModel(
-			dataProviderInstanceElement,
-			ExportImportPathUtil.getModelPath(dataProviderInstance),
-			dataProviderInstance);
+		throws PortletDataException {
 	}
 
 	@Override
-	protected void doImportStagedModel(
+	public DDMDataProviderInstance saveStagedModel(
+			DDMDataProviderInstance dataProviderInstance)
+		throws PortalException {
+
+		return _ddmDataProviderInstanceLocalService.
+			updateDDMDataProviderInstance(dataProviderInstance);
+	}
+
+	@Override
+	public DDMDataProviderInstance updateStagedModel(
 			PortletDataContext portletDataContext,
 			DDMDataProviderInstance dataProviderInstance)
-		throws Exception {
-
-		DDMDataProvider ddmDataProvider =
-			_ddmDataProviderTracker.getDDMDataProvider(
-				dataProviderInstance.getType());
-
-		if (ddmDataProvider == null) {
-			throw new IllegalStateException(
-				"No such DataProvider of type " + dataProviderInstance.getType()
-			);
-		}
+		throws PortalException {
 
 		long userId = portletDataContext.getUserId(
 			dataProviderInstance.getUserUuid());
-
-		DDMDataProviderInstance importedProvider =
-			(DDMDataProviderInstance)dataProviderInstance.clone();
-
-		importedProvider.setGroupId(portletDataContext.getScopeGroupId());
-
-		DDMDataProviderInstance existingProvider =
-			_ddmDataProviderInstanceLocalService.
-				fetchDDMDataProviderInstanceByUuidAndGroupId(
-					dataProviderInstance.getUuid(),
-					portletDataContext.getScopeGroupId());
 
 		ServiceContext serviceContext = portletDataContext.createServiceContext(
 			dataProviderInstance);
@@ -159,38 +172,16 @@ public class DDMDataProviderInstanceStagedModelDataHandler
 		DDMFormValues ddmFormValues =
 			_ddmFormValuesJSONDeserializer.deserialize(ddmForm, definition);
 
-		if (existingProvider == null) {
-			serviceContext.setUuid(dataProviderInstance.getUuid());
-
-			importedProvider =
-				_ddmDataProviderInstanceLocalService.addDataProviderInstance(
-					userId, portletDataContext.getScopeGroupId(),
-					dataProviderInstance.getNameMap(),
-					dataProviderInstance.getDescriptionMap(), ddmFormValues,
-					dataProviderInstance.getType(), serviceContext);
-		}
-		else {
-			importedProvider.setDataProviderInstanceId(
-				existingProvider.getDataProviderInstanceId());
-
-			importedProvider =
-				_ddmDataProviderInstanceLocalService.updateDataProviderInstance(
-					userId, existingProvider.getDataProviderInstanceId(),
-					dataProviderInstance.getNameMap(),
-					dataProviderInstance.getDescriptionMap(), ddmFormValues,
-					serviceContext);
-		}
-
-		portletDataContext.importClassedModel(
-			dataProviderInstance, importedProvider);
+		return _ddmDataProviderInstanceLocalService.updateDataProviderInstance(
+			userId, dataProviderInstance.getDataProviderInstanceId(),
+			dataProviderInstance.getNameMap(),
+			dataProviderInstance.getDescriptionMap(), ddmFormValues,
+			serviceContext);
 	}
 
 	@Reference
 	private DDMDataProviderInstanceLocalService
 		_ddmDataProviderInstanceLocalService;
-
-	@Reference
-	private DDMDataProviderTracker _ddmDataProviderTracker;
 
 	@Reference
 	private DDMFormValuesJSONDeserializer _ddmFormValuesJSONDeserializer;
